CRAB-34229:move methods util to resolve cyclic dependencies
