+++ app/ami.py
@@ -3,9 +3,7 @@ from dataclasses import dataclass
 from app import parsing
 from app import util
 from .resource import Resource
-from .snapshot import estimate_monthly_snapshot_price
 import boto3
-import botocore
 
 
 @dataclass
@@ -191,19 +189,9 @@ def estimate_monthly_ami_price(ami_type: str, block_device_mappings: list, ami_n
                 snapshot = device["Ebs"]
                 snapshot_type = snapshot["VolumeType"]
                 snapshot_size = snapshot["VolumeSize"]
-                total_cost += estimate_monthly_snapshot_price(snapshot_type, snapshot_size)
+                total_cost += util.estimate_monthly_snapshot_price(snapshot_type, snapshot_size)
     else:
         print(f"WARNING: {ami_name} is a {ami_type} type AMI with the following block_device_mappings: "
               f"{block_device_mappings}")
     return total_cost
 
-
-def is_ami_registered(ami_id: str, region_name: str) -> bool:
-    ec2 = boto3.resource('ec2', region_name=region_name)
-    is_registered = True
-    # Retrieve name of AMI, if ClientError or AttributeError is thrown, the AMI does not exist
-    try:
-        ec2.Image(ami_id).name
-    except (botocore.exceptions.ClientError, AttributeError):
-        is_registered = False
-    return is_registered
