+++ app/snapshot.py
@@ -1,12 +1,10 @@
 from dataclasses import dataclass
 
 from app import parsing
-from app import ami
 from app import util
 from .resource import Resource
 
 import boto3
-import re
 
 
 @dataclass
@@ -83,11 +81,11 @@ class Snapshot(Resource):
         resource_id_tag = 'SnapshotId'
         resource_type_tag = 'StorageTier'
 
-        monthly_price = estimate_monthly_snapshot_price(snapshot_type, size)
+        monthly_price = util.estimate_monthly_snapshot_price(snapshot_type, size)
 
         snapshot = Resource.build_generic_model(tags, resource_dict, region_name, resource_id_tag, resource_type_tag)
         is_aws_backup_snapshot, is_ami_snapshot = \
-            is_backup_or_ami_snapshot(resource_dict['Description'], region_name)
+            util.is_backup_or_ami_snapshot(resource_dict['Description'], region_name)
 
         return Snapshot(region_name=region_name,
                         resource_id=snapshot.resource_id,
@@ -153,27 +151,3 @@ class Snapshot(Resource):
             return True
         else:
             return False
-
-
-# Estimated monthly costs were formulated by taking the average monthly costs of N. California and Oregon
-def estimate_monthly_snapshot_price(type: str, size: float) -> float:
-    standard_monthly_cost = .0525
-    archive_monthly_cost = .0131
-    return standard_monthly_cost*size if type == "standard" else archive_monthly_cost*size
-
-
-# Checks the snapshot description to see if the snapshot is part of an AMI or AWS backup.
-# If the snapshot is part of an AMI, but the AMI has been deregistered, then this function will return False
-# for is_ami_snapshot so the remaining snapshot can be cleaned up.
-def is_backup_or_ami_snapshot(description: str, region_name: str) -> bool:
-    is_aws_backup_snapshot = False
-    is_ami_snapshot = False
-    if "AWS Backup service" in description:
-        is_aws_backup_snapshot = True
-    elif "Copied for DestinationAmi" in description:
-        # regex matches the first occurrence of ami, since the snapshot
-        # belongs to the first mentioned ami (destination ami) and not the second (source ami)
-        ami_id = re.search(r'ami-\S*', description).group()
-        is_ami_snapshot = ami.is_ami_registered(ami_id, region_name)
-
-    return is_aws_backup_snapshot, is_ami_snapshot
